Tests/LibWeb/Text/input/cancel-of-pointerdown-or-mousedown-should-skip-focus-steps.html

<!DOCTYPE html>
<button id="test-button" style="width: 100px; height: 100px">Test button</button>
<script src="include.js"></script>
<script>
    asyncTest(done => {
        const button = document.getElementById("test-button");

        const eventsToLog = [
            "mousedown",
            "mouseup",
            "click",
            "dblclick",
            "pointerdown",
            "pointerup",
            "focus",
            "blur",
            "focusin",
            "focusout",
        ];

        let eventToPrevent;
        eventsToLog.forEach(type => {
            button.addEventListener(type, event => {
                if (type === eventToPrevent) event.preventDefault();
                println(`${type}`, {
                    key: event.key,
                    code: event.code,
                    button: event.button,
                    pointerType: event.pointerType,
                    eventPhase: event.eventPhase,
                    timeStamp: event.timeStamp,
                });
            });
        });

        println("> preventing pointerdown");
        eventToPrevent = "pointerdown";
        internals.click(50, 50);

        println("> preventing mousedown");
        eventToPrevent = "mousedown";
        internals.click(50, 50);

        requestIdleCallback(() => done());
    });
</script>
