Tests/LibWeb/Text/expected/cancel-of-pointerdown-or-mousedown-should-skip-focus-steps.txt

> preventing pointerdown
pointerdown
pointerup
mouseup
click
> preventing mousedown
pointerdown
mousedown
pointerup
mouseup
click
